News_AI_Agent/pyproject.toml
[project]
name = "news-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "crewai",
    "crewai-tools",
    "langchain-google-genai",
    "load-dotenv",
]
